@@ -1,9 +1,10 @@
-const axios = require('axios');
+import axios from 'axios';
 
-export const getDemo = params => {
-	const URL = 'https://reactjsteachingproj.herokuapp.com/users';
+const baseURL = 'https://www.omdbapi.com/?apiKey=ffd0c3a5';
 
-	return axios.get(URL).then(response => {
-		return response.data;
-	});
-};
+export const apiCall = (url, data, headers, method) => axios({
+    method,
+    url: baseURL + url,
+    data,
+    headers
+});
