@@ -1,3 +1,7 @@
-export const GET_DEMO_REQUEST = 'GET_DEMO_REQUEST';
-export const GET_DEMO_SUCCESS = 'GET_DEMO_SUCCESS';
-export const SEARCH_MEDIA_ERROR = 'SEARCH_MEDIA_ERROR';
+export const SEARCH_MOVIE_START = 'SEARCH_MOVIE_START';
+export const SEARCH_MOVIE_ERROR = 'SEARCH_MOVIE_ERROR';
+export const SEARCH_MOVIE_COMPLETE = 'SEARCH_MOVIE_COMPLETE';
+
+export const SEARCH_MOVIE_BY_ID_START = 'SEARCH_MOVIE_BY_ID_START';
+export const SEARCH_MOVIE_BY_ID_ERROR = 'SEARCH_MOVIE_BY_ID_ERROR';
+export const SEARCH_MOVIE_BY_ID_COMPLETE = 'SEARCH_MOVIE_BY_ID_COMPLETE';
